splitting functions to more files, now the norm is ok, however ambient light stopped working...
