+++ .gitignore
@@ -2,3 +2,4 @@
 minirt
 *.kdev4
 .kdev4/
+e2e.log
