+++ src/debug/scene_print.c
@@ -6,7 +6,7 @@
 /*   By: smelicha <smelicha@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/03 18:49:51 by smelicha          #+#    #+#             */
-/*   Updated: 2025/05/03 18:51:26 by smelicha         ###   ########.fr       */
+/*   Updated: 2025/05/03 19:54:40 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
