+++ src/parser/scene_parser.c
@@ -6,58 +6,11 @@
 /*   By: smelicha <smelicha@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 15:57:23 by jstudnic          #+#    #+#             */
-/*   Updated: 2025/05/03 14:10:10 by smelicha         ###   ########.fr       */
+/*   Updated: 2025/05/03 22:21:52 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incl/minirt.h"
-// #include <type_traits>
-
-t_rgb	parse_color(char *str)
-{
-	t_rgb	color;
-	char	**values;
-
-	values = ft_split(str, ',');
-	if (!values || double_array_length(values) != 3)
-	{
-		if (values)
-			ft_free_split(values);
-		printf("parse colors returning invalid because of number of tokens\n");
-		return ((t_rgb){-1, -1, -1});
-	}
-	color = create_rgb(ft_atoi(values[0]), \
-		ft_atoi(values[1]), \
-		ft_atoi(values[2]));
-	ft_free_split(values);
-	if (!validate_rgb(color))
-	{
-		printf("parse colors returning invalid because of invalid colors\n");
-		return ((t_rgb){-1, -1, -1});
-	}
-	return (color);
-}
-
-t_vector	parse_vector(char *str)
-{
-	t_vector	vec;
-	char		**values;
-
-	printf("parse vector: string to parse: %s\n", str);
-	values = ft_split(str, ',');
-	if (!values || double_array_length(values) != 3)
-	{
-		if (values)
-			ft_free_split(values);
-		printf("parse vector returning because of wrong number of tokens\n");
-		return ((t_vector){NAN, NAN, NAN});
-	}
-	vec.x = ft_atof(values[0]);
-	vec.y = ft_atof(values[1]);
-	vec.z = ft_atof(values[2]);
-	ft_free_split(values);
-	return (vec);
-}
 
 static int	parse_ambient(char *line, t_scene *scene)
 {
@@ -101,29 +54,6 @@ static int	parse_camera(char *line, t_scene *scene)
 		|| scene->camera.fov <= 0 || scene->camera.fov >= 180)
 		return (0);
 	return (1);
-}
-
-static int	parse_light(char *line, t_scene *scene)
-{
-	char	**tokens;
-	t_light	light;
-
-	tokens = ft_split(line, ' ');
-	if (!tokens || double_array_length(tokens) != 4)
-	{
-		if (tokens)
-			ft_free_split(tokens);
-		printf("parse light returning because of wrong number of tokens\n");
-		return (0);
-	}
-	light.position = parse_vector(tokens[1]);
-	light.brightness = ft_atof(tokens[2]);
-	light.color = parse_color(tokens[3]);
-	ft_free_split(tokens);
-	if (isnan(light.position.x) || light.color.r == -1
-		|| light.brightness < 0.0 || light.brightness > 1.0)
-		return (0);
-	return (add_light(scene, light));
 }
 
 typedef struct s_ps
