+++ src/render/vectors.c
@@ -6,7 +6,7 @@
 /*   By: smelicha <smelicha@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/06 14:16:26 by jstudnic          #+#    #+#             */
-/*   Updated: 2025/05/03 13:42:50 by smelicha         ###   ########.fr       */
+/*   Updated: 2025/05/03 22:35:12 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
