+++ src/utils/ft_atoi_atof.c
@@ -6,7 +6,7 @@
 /*   By: smelicha <smelicha@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 16:44:23 by jstudnic          #+#    #+#             */
-/*   Updated: 2025/05/03 16:53:51 by smelicha         ###   ########.fr       */
+/*   Updated: 2025/05/03 21:58:48 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
