+++ src/utils/get_next_line.c
@@ -6,77 +6,13 @@
 /*   By: smelicha <smelicha@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 16:48:23 by jstudnic          #+#    #+#             */
-/*   Updated: 2025/05/03 19:00:57 by smelicha         ###   ########.fr       */
+/*   Updated: 2025/05/03 22:08:37 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incl/libft.h"
 
 #define BUFFER_SIZE 1024
-
-static char	*ft_strchr(const char *s, int c)
-{
-	while (*s)
-	{
-		if (*s == (char)c)
-			return ((char *)s);
-		s++;
-	}
-	if ((char)c == '\0')
-		return ((char *)s);
-	return (NULL);
-}
-
-static char	*ft_strjoin(char const *s1, char const *s2)
-{
-	char	*result;
-	size_t	i;
-	size_t	j;
-	size_t	len1;
-	size_t	len2;
-
-	if (!s1 || !s2)
-		return (NULL);
-	len1 = 0;
-	while (s1[len1])
-		len1++;
-	len2 = 0;
-	while (s2[len2])
-		len2++;
-	result = malloc(sizeof(char) * (len1 + len2 + 1));
-	if (!result)
-		return (NULL);
-	i = -1;
-	while (++i < len1)
-		result[i] = s1[i];
-	j = -1;
-	while (++j < len2)
-		result[i + j] = s2[j];
-	result[i + j] = '\0';
-	return (result);
-}
-
-static char	*ft_strdup(const char *s1)
-{
-	char	*result;
-	size_t	len;
-	size_t	i;
-
-	len = 0;
-	while (s1[len])
-		len++;
-	result = malloc(sizeof(char) * (len + 1));
-	if (!result)
-		return (NULL);
-	i = 0;
-	while (i < len)
-	{
-		result[i] = s1[i];
-		i++;
-	}
-	result[i] = '\0';
-	return (result);
-}
 
 static char	*read_line(int fd, char *buffer, char *backup)
 {
